+++ lib/src/state/polygons.dart
@@ -1,6 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
+import 'package:geojson/geojson.dart';
+import 'package:geopoint/geopoint.dart';
 import 'package:latlong2/latlong.dart';
+
+import '../controller.dart';
 import '../models.dart';
 
 /// State of the polygons on the map
@@ -9,7 +13,7 @@ class PolygonsState {
   PolygonsState({required this.notify});
 
   /// The notify function
-  final Function notify;
+  final FeedNotifyFunction notify;
 
   final Map<String, Polygon> _namedPolygons = {};
 
@@ -20,12 +24,13 @@ class PolygonsState {
   List<Polygon> get polygons => _namedPolygons.values.toList();
 
   /// Add a polygon on the map
-  Future<void> addPolygon(
-      {required String name,
-      required List<LatLng> points,
-      required Color color,
-      required double borderWidth,
-      required Color borderColor}) async {
+  void addPolygon({
+    required String name,
+    required List<LatLng> points,
+    required Color color,
+    required double borderWidth,
+    required Color borderColor,
+  }) {
     _namedPolygons[name] = Polygon(
         points: points,
         color: color,
@@ -36,11 +41,45 @@ class PolygonsState {
   }
 
   /// Remove a polygon from the map
-  Future<void> removePolygon(String name) async {
+  void removePolygon(String name) {
     if (_namedPolygons.containsKey(name)) {
       _namedPolygons.remove(name);
       notify("updatePolygons", name, removePolygon,
           MapControllerChangeType.polygons);
     }
+  }
+
+  /// Remove multiple polygons from the map
+  void removePolygons(List<String> names) {
+    // TODO: Optimize this
+    for (final name in names) {
+      removePolygon(name);
+    }
+  }
+
+  /// Export all polygons to a [GeoJsonFeature] with geometry
+  /// type [GeoJsonMultiPolygon]
+  GeoJsonFeature<GeoJsonMultiPolygon>? toGeoJsonFeatures() {
+    if (namedPolygons.isEmpty) {
+      return null;
+    }
+    final multiPolygon = GeoJsonMultiPolygon(name: "map_polygons");
+    for (final k in namedPolygons.keys) {
+      final mapPolygon = namedPolygons[k]!;
+      final polygon = GeoJsonPolygon()..name = k;
+      final geoSerie =
+          GeoSerie(name: polygon.name!, type: GeoSerieType.polygon);
+      for (final point in mapPolygon.points) {
+        geoSerie.geoPoints.add(
+            GeoPoint(latitude: point.latitude, longitude: point.longitude));
+      }
+      polygon.geoSeries = [geoSerie];
+      multiPolygon.polygons.add(polygon);
+    }
+    final feature = GeoJsonFeature<GeoJsonMultiPolygon>()
+      ..type = GeoJsonFeatureType.multipolygon
+      ..geometry = multiPolygon
+      ..properties = <String, dynamic>{"name": "map_polygons"};
+    return feature;
   }
 }